Iso time for catalog and workflow up5631 (#246)
* roll over iso time period tests to utils from workflow

* adjust in workflow

* format_time_period in utils

* also use iso time format in catalog
