+++ up42/workflow.py
@@ -6,7 +6,6 @@ from pathlib import Path
 from typing import Dict, List, Union, Tuple, Optional
 import warnings
 from datetime import datetime
-from datetime import time as datetime_time
 
 from geopandas import GeoDataFrame
 from shapely.geometry import Point, Polygon
@@ -25,6 +24,7 @@ from up42.utils import (
     any_vector_to_fc,
     fc_to_query_geometry,
     filter_jobs_on_mode,
+    format_time_period,
 )
 
 logger = get_logger(__name__)
@@ -383,41 +383,10 @@ class Workflow:
                 input_parameters[data_block_name]["limit"] = len(scene_ids)
                 input_parameters[data_block_name].pop("time")
             elif start_date is not None or end_date is not None:
-                if start_date is None or end_date is None:
-                    raise ValueError(
-                        "When using dates, both start_date and end_date need to be provided."
-                    )
-                # Start and end date can be any combination of str ("YYYY-MM-DD" or "YYYY-MM-DDTHH:MM:SS")
-                # or datetime objects.
-                if not isinstance(start_date, datetime):
-                    start_dt: datetime = datetime.fromisoformat(start_date)
-                else:
-                    start_dt = start_date
-
-                if not isinstance(end_date, datetime):
-                    end_dt: datetime = datetime.fromisoformat(end_date)
-                    try:
-                        # For "YYYY-MM-DD" string the default datetime conversion sets to
-                        # start of day, but image archive query requires end of day.
-                        datetime.strptime(end_date, "%Y-%m-%d")  # format validation
-                        end_dt = datetime.combine(
-                            end_dt.date(), datetime_time(23, 59, 59, 999999)
-                        )
-                    except ValueError:
-                        pass
-                else:
-                    end_dt = end_date
-
-                if start_dt > end_dt:
-                    raise ValueError(
-                        "The start_date can not be later than the end_date!"
-                    )
-
-                formatting = "%Y-%m-%dT%H:%M:%S"
-                period = (
-                    f"{start_dt.strftime(formatting)}Z/{end_dt.strftime(formatting)}Z"
+                time_period = format_time_period(
+                    start_date=start_date, end_date=end_date
                 )
-                input_parameters[data_block_name]["time"] = period
+                input_parameters[data_block_name]["time"] = time_period
 
             if geometry is not None:
                 aoi_fc = any_vector_to_fc(
